requirements.txt
# Python runtime used in the current working environment:
# Python 3.10.11

# Core numeric/audio
numpy==1.22.0
sounddevice==0.5.5
soundfile==0.13.1
pygame==2.6.1

# ASR / VAD / denoise
sherpa-onnx==1.12.32
torch==2.10.0
torchaudio==2.10.0
pyrnnoise==0.4.3
speechbrain>=1.0.0,<2.0.0

# MT API and client
fastapi==0.135.1
uvicorn[standard]==0.42.0
pydantic==2.12.5
requests==2.32.5
transformers==4.41.0
sentencepiece==0.2.1

# TTS
TTS==0.22.0
edge-tts==7.2.8
dashscope==1.25.14

# Infrastructure
redis==7.3.0
python-dotenv==1.2.2
pywebview==5.1
